Writing unit tests
Write unit tests, the unit tests will create a new dummy connection and
will let the test function to execute commands directly.

This is the first and fast tests, for sanity checks. The ultimate goal
is to run the tcl tests from the redis projects as we are aiming to be
compatible.

Also improved issued found while running the unit tests. For instance
the way returned data are cloned is improved. It is now harder to clone
complex values (Hashes, set). It is not longer possible to read complex
values through `DB.get`, `DB.get_map_or` should be used instead.