Use Bytes instead of checksum::Value for Set.
checksum::Value is only useful when comparing items one by one, like in
linked lists. Sets are hashed, so it makes no sense to use
checksum::Value, instead Bytes are used.

This change will reduce the size of stored items in the set and will
avoid cloning data to query the set.