Introducing Entry::Version
Entry::Version is the nano-second of the last update (or creation) of
the entry. This is needed to keep track of key changes for transaction
support.

When reading values through `get_map_or` be sure to bump the version
(calling `db.bump_version`) outside of the callbacks to avoid deadlocks.

Check the documentation for WATCH/UNWATCH/EXEC commands.