Added support for multi/exec.
This implementation will not lock the entire database to process one
transaction at a time, as the original implementation of Redis does,
instead the list of keys affected in the transaction are locked
exclusively by the transaction. Any other connection trying to read or
update locked keys will have to wait in-line.

This implementation is multi-threaded and async I/O, that is why I don't
want to lock-up the entire database exclusively for one transaction,
instead I'm trying to keep the same behaviour as Redis does by locking
all keys involved exclusively.

TODO:
* [ ] Maybe create the loop/await async, instead of blocking the current
      thread.