Add support for Transactions (#9)
* Introducing Entry::Version

Entry::Version is the nano-second of the last update (or creation) of
the entry. This is needed to keep track of key changes for transaction
support.

When reading values through `get_map_or` be sure to bump the version
(calling `db.bump_version`) outside of the callbacks to avoid deadlocks.

Check the documentation for WATCH/UNWATCH/EXEC commands.

* Added support for `watch` and `unwatch`

This command will keep a list of watched keys and their current version.
The `unwatch` command will clean up that list.

The list will be used by `multi`/`exec` to skip a transaction if the
watched keys changed after being watched.

* Added code to extract keys from commands

The idea is to keep track of which keys are passed in each command, this
will help keep track of which keys should be locked exclusively when a
transaction is executed.

* Added function to keep track of keys

* Adding code to start, exec and discard transactions

* Remove deref() call

* Added support for multi/exec.

This implementation will not lock the entire database to process one
transaction at a time, as the original implementation of Redis does,
instead the list of keys affected in the transaction are locked
exclusively by the transaction. Any other connection trying to read or
update locked keys will have to wait in-line.

This implementation is multi-threaded and async I/O, that is why I don't
want to lock-up the entire database exclusively for one transaction,
instead I'm trying to keep the same behaviour as Redis does by locking
all keys involved exclusively.

TODO:
* [ ] Maybe create the loop/await async, instead of blocking the current
      thread.