Prototype: Added pub-sub
The entire pub-sub is quite simple and rely heavely on Tokio's mpsc
channels.

The logic to subscribe to patterns is not yet done, but it will be added
in the future with heavy caching to make sure it is as efficient as it
can possibly be.

The main evented-loop will read off the client and from the internal
pub-sub channels to send a response.