Fixed bug with PERSIST
PERSIST operation did not remove the key from the purge process.
Microredis does have a dedicated structure to watch keys with
expirations in it.

Previous this commit if a key had any expiration and then this
expiration was removed with the PERSIST command, the expiration would be
removed from the key-itself, but the purge process (which is a
background process) wouldn't know about this change removing the key in
due time.

This commit notifies of PERSIST changes to the purge process.