Add multiple databases support (#32)
* Minor improvements

Minor improvements needed to make it easier to add support for multiple
databases.

* Add databases pool

This new struct is going to host all databases in the server. Each
database will have their own Db, to keep the codebase simple.

New methods are going to be added to move and copy databases from
databases atomically.