Added `keys` command (#14)
* Added `keys` command

Added `keys` command.

* Add TYPE command

* Added UNLINK which is an alias to DEL

* Added COPY command

* Added support for MOVE

Also renamed all command definitions to uppercase to avoid using
reserved keywords.

* Added RENAME and RENAMENX

* Added OBJECT command handler

Also improved the help sector adding a cmd/help.rs which hosts all the
help texts for all commands.

* Added Cursor type (#35)

The cursor type is serialized an string with integers. The integer
contains a tuple with (bucket_id, last_position) and a checksum. If the
cursor is invalid a new cursor with position (0, 0) is deserialized
instead.