Add support to Redis test suite (#41)
* Fixed config and unix socket issues

* Added option to parse Option<t>.
* Fixed issue with unix socket. If file exists it must be removed.

* Update redis protocol parser lib

Updated library which added support for inline command parsing. Redis'
test suite uses inline command, so the support was a hard dependency.

Added also some logging information expected by the test suite,
specially to know when the server is ready to accept connections.

* Add FLUSHDB and DBSIZE commands

These commands are needed for test suite.

* Add INFO and improve SCAN parsing

Added INFO command and improve SCAN argument parsing to comply with the
main implementation.

* Add DEBUG command

* Added DEBUG command
* Fixed few internal typos in the Db mod
* Add Value digest()

* Add support for object debug

* Fixed bug in RENAMENX found with integrational test

* Fixed SETNX bug found by test suit

Fetching the previous would select event expired (but not yet collected)
data.

* Fix bug in GETEX PERSIST found by test suite

* Fixed bug in MSET / MSETNX

The Database was not checking if the list of key,value was complete or
not. Leading to a potential panic.

The issue was found by Redis test suite.

The fix also return the expected error.

* Fixed SETRANGE error handling

* Handle negative offset with an error (instead of invalid number)
* Handle max size

* Fixed bug with SET argument parsing

If the `GET` argument is passed, it should return the previous value or
Null.

This issue has been found with the redis test suite.

* Fixed GETRANGE math panic

If the value's len is 0 it would panic. This issue has been found by
redis' test suite.

* Fixed SET .. .. GET error

If the previous value is not a stream of bytes the whole operation must
fail and no changes should happen.

* Change error for INCR for non-string types

Instead of returning an NotANumber error it should return a WrongType
error

* Improve INCR* casting

Float numbers will be converted to integers if it is possible without
any loss.

* Rename error to be more standard with Redis

* Added support for negative RANK in LPOS

* Improved LINDEX

* Added better support for negative offsets
* Not found keys should return Null instead of `0`

* Improve `remove_element`

It was receiving usize instead of Option<usize> and using `0` instead of
None and it was very confusing.

* Fixed BLPOP/BRPOP timeout=0

It should wait forever, instead of giving up right away.

* Update `remove_element`

Delete key if the list is empty after removing elements

* Improve error code

* Update BLPOP/BRPOP timeout parsing errors

* Working on `client unblock`

* Improved LRANGE

* Change error while trying to parse int instead of number

* Improved `smismember` when key does not exists

* Fixed bug for SUNION when first key does not exists

* Improve sdiffstore

sdiffstore should delete the key if the sdiff is empty.

* Improved SINTER with empty keys

* Improved sinterstore and sunionstore

* SPOP/SRANDMEMBER must return Null when the key is not found instead of 0

* Add support for srandmember with negative limits

* Improved SMOVE

* Removed deadlock when the source/target are the same key
* Check if the key is the a set *before* doing any other check
* Improve other error cases

* Fixed response of HMGET when key doesn't exists

* Improve HDEL

If hash is empty after HDEL the key must be deleted

* Improve INCR / HINCRBY / HINCRBYFLOAT

* Improved internal storage. Rounding to an integer if possible
* Move HINCRY* to the Database for better edge case handling

* Minor changes

* Added Overflow protection for INCR/HINCR

The overflow is for f64 and integers

* Float must be returned as strings

In INCR/HINCR

* Add support for failed tx

Failed Tx will continue queueing commands as regular but won't execute
and will return a very particular error.

* SPOP should remove key if the set is empty

* Improved WATCH behaviour

WATCH calls are not allowed inside Multi, only before.

* Added support for FLUSHALL

* Fix WATCH for non-existing keys

It should be watch a `0` version instead of current.

* Implement QUIT command

* Improvements

 * Fixed PING to be called inside pubsub
 * Changed error for invalid expiration

* Fixed expiration overflows

* Adding extra params to EXPIRE / EXPIREAT

* Parse EXPIRE/EXPIREAT options

* Fixed time parsing

* Make sure exists only counts non-expired keys

A key may exists, because it was not yet gargabe collected. Before
counting it must be checked to be valid.

* Improved timestamp parsing and its errors

* Improved data mutation

* Improved PING in pubsub mode

* Improved SUBSCRIBE support

* PING support
* Added Value::Ignore (this value is not returned to the client).
* Fixed PUBSUB NUMPAT

* Add Redis tests

Added official redis test suite from their repo. A few tests were
disabled because they are not yet supported or they won't be ever
supported.

The goal going forward is to keep updating the tests regularly.

Credits: https://github.com/redis/redis